requirements.txt
aiohttp==3.7.3
aiohttp-jwt==0.6.1
alembic==1.8.0
async-timeout==3.0.1
asyncpg==0.25.0
attrs==21.4.0
certifi==2022.6.15
chardet==3.0.4
charset-normalizer==2.1.0
et-xmlfile==1.1.0
gino==1.0.1
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.8.0
Mako==1.2.1
MarkupSafe==2.1.1
multidict==6.0.2
numpy==1.23.0
openpyxl==3.0.10
psycopg2==2.9.3
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
six==1.16.0
SQLAlchemy==1.3.24
typing-extensions==4.3.0
urllib3==1.26.9
xlrd==1.2.0
yarl==1.7.2
zipp==3.8.0
